src/auth-service/requirements.txt
flask
flask-sqlalchemy
gunicorn
sqlalchemy
psycopg2-binary
alembic
pydantic
omegaconf
redis
bcrypt
pyjwt[crypto]
cryptography
python-dotenv
flask-cors
email-validator
pytest
pytest-cov
pytest-mock